# src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "maklaGo API",
    "version": "1.0.0",
    "description": "Restaurant POS backend. **Tenant**: optional header `x-tenant-id` (slug or registry UUID). With `Authorization: Bearer` staff JWT, the tenant is resolved from the token when the header is omitted. **Language**: optional header **`lang`** (`fr`, `en`, `ar`; default **`fr`**) on every request — error messages and localized strings follow it; response includes **`Content-Language`**. **`POST /api/v1/auth/login`** accepts optional `x-tenant-id` for tenant-scoped login; without it, the normalized email must exist in registry **`staff_login_directory`** (slugless directory login). **`GET /api/v1/orders`** returns enriched catalog labels on lines (dashboard); **`GET /api/v1/mobile/orders`** uses a slimmer shell than **`POST /orders`**: no **`staff`**, no nested **`table`** object; dine-in adds root **`tableId`** + **`tableNumber`**. Lines add **`products[].name`**, **`extras[].name`**, optional **`extras[].typeId`**, optional **`compositionSnapshot`**. **`POST /api/v1/orders`** accepts **`products`** snapshots (`categoryId`, `id`, `count`, **`price`**, **`extras`**) for analytics; **`subtotal`**, **`tva`**, **`total`** use major currency units in JSON (e.g. 2.5); DB keeps cents internally (must match server); new orders start as **`waiting`**. **Platform** routes use `x-platform-key`. Errors: `{ \"error\": { \"code\", \"message\", \"details?\" } }`."
  },
  "servers": [
    {
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Staff JWT from `/api/v1/auth/login`."
      },
      "PlatformKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-platform-key",
        "description": "Must match `PLATFORM_API_KEY` on the server."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      },
...
      "OrderStatus": {
        "type": "string",
        "enum": [
          "waiting",
          "confirmed",
          "preparing",
          "completed",
          "canceled"
        ]
      },
...
      "PatchOrderTableRequest": {
        "type": "object",
        "required": ["tableId"],
        "properties": {
          "tableId": {
            "type": "string",
            "minLength": 12,
            "maxLength": 36,
            "description": "Active restaurant table id for this dine-in order."
          }
        }
      },
...
      "OrderSerializedEnriched": {
        "type": "object",
        "description": "Backoffice order (`GET /orders`, `GET /orders/{id}`, `PATCH …/status` HTTP body): enriched **`products`** (`OrderProductEnriched`). Shell matches **`OrderSerialized`** (table rules, `staff`, etc.).",
        "properties": {
          "id": { "type": "string", "minLength": 12, "maxLength": 36 },
          "reference": {
            "type": "string",
            "description": "Unique ticket reference: **8** lowercase hexadecimal characters for newly created orders.",
            "example": "a3f9c21d"
          },
          "commandNumber": { "type": "integer" },
          "commandDate": { "type": "string", "format": "date" },
          "status": { "$ref": "#/components/schemas/OrderStatus" },
          "orderType": { "type": "string", "enum": ["dine_in", "takeaway"] },
          "tableId": {
            "type": "string",
            "minLength": 12,
            "maxLength": 36,
            "description": "Present only for dine-in orders with an assigned table."
          },
          "staffId": { "type": "string", "minLength": 12, "maxLength": 36 },
          "note": { "type": "string", "nullable": true },
          "customerName": { "type": "string", "nullable": true },
          "discount": { "type": "number", "description": "Order-level discount percent (0–100)." },
          "discountPrice": { "type": "number" },
          "paymentMethod": { "type": "string", "enum": ["cash", "card", "unpaid"] },
          "subtotal": { "type": "number", "description": "Sum of line net totals (major currency units)." },
          "tva": { "type": "number", "description": "Sum of line TVA amounts (major currency units)." },
          "total": { "type": "number", "description": "Final total after discount (major currency units)." },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "products": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrderProductEnriched" }
          },
          "table": {
            "$ref": "#/components/schemas/OrderTableSerialized",
            "description": "Present only with `tableId` on dine-in orders when table row was loaded.",
            "nullable": true
          },
          "staff": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          }
        }
      },
      "TableBody": {
        "type": "object",
        "required": [
          "tableNumber"
        ],
        "properties": {
          "tableNumber": {
            "type": "integer",
            "minimum": 1,
            "description": "Numeric label; **unique together with `zoneId`** (different zones may reuse the same number)."
          },
          "zoneId": {
            "type": "string",
            "minLength": 12,
            "maxLength": 36,
            "description": "**POST /tables**: required — table numbers are unique **per zone**. **PATCH**: optional; omit keys you are not updating."
          },
          "seatCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 999,
            "default": 4
          },
          "sortOrder": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": ["free", "occupied"],
            "description": "Patch only; normally derived from open dine-in orders."
          }
        }
      },
      "TableZone": {
        "type": "object",
        "description": "Dining area / floor section. **`RestaurantTable.zoneId`** references this; delete is blocked while tables still reference the zone (`409 zone_has_tables`). Zone **`name`** is unique.",
        "required": ["id", "name", "sortOrder", "createdAt"],
        "properties": {
          "id": { "type": "string", "minLength": 12, "maxLength": 36 },
          "name": { "type": "string", "maxLength": 100 },
          "sortOrder": { "type": "integer", "description": "Lower sorts first in listings." },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "TableZoneCreateBody": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "sortOrder": { "type": "integer", "description": "Defaults to **0** when omitted." }
        }
      },
      "TableZonePatchBody": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "sortOrder": { "type": "integer" }
        }
      },
...
      "TableZoneWithTablesMobile": {
        "type": "object",
        "required": ["id", "name", "sortOrder", "createdAt", "tables"],
        "properties": {
          "id": { "type": "string", "minLength": 12, "maxLength": 36 },
          "name": { "type": "string" },
          "sortOrder": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "tables": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MobileTableInZone" }
          }
        }
      },
...
      "StatsFilter": {
        "type": "string",
        "enum": ["today", "week", "month", "year", "custom"],
        "default": "today"
      },
...
      "OrderStatusCounts": {
        "type": "object",
        "properties": {
          "waiting": { "type": "integer" },
          "confirmed": { "type": "integer" },
          "preparing": { "type": "integer" },
          "completed": { "type": "integer" },
          "canceled": { "type": "integer" }
        }
      },
      "StatisticsResponse": {
        "type": "object",
        "description": "Dashboard statistics (amounts in major currency units, e.g. 2.5 TND). Period boundaries use Africa/Tunis.",
        "properties": {
          "totalRevenue": { "type": "number" },
          "completedOrders": { "type": "integer", "description": "Revenue-eligible order count (non-canceled) in period" },
          "moyenRevenue": { "type": "number", "description": "Average order value in period" },
          "paymentMethodsTotalRevenue": {
            "type": "object",
            "properties": {
              "cash": { "type": "number" },
              "card": { "type": "number" },
              "cashCount": { "type": "integer" },
              "cardCount": { "type": "integer" }
            }
          },
          "orderTypes": {
            "type": "object",
            "description": "Dine-in vs takeaway breakdown",
            "properties": {
              "dineInCount": { "type": "integer" },
              "takeawayCount": { "type": "integer" },
              "dineIn": { "type": "number" },
              "takeaway": { "type": "number" }
            }
          },
          "totalOrders": { "type": "integer" },
          "totalPlat": { "type": "integer", "description": "All-time units sold (non-canceled orders)" },
          "orderStatuses": { "$ref": "#/components/schemas/OrderStatusCounts" },
          "revenueComparison": {
            "type": "object",
            "properties": {
              "currentRevenue": { "type": "number" },
              "previousRevenue": { "type": "number" },
              "difference": { "type": "number" },
              "percentageChange": { "type": "number" },
              "currentRevenuePercentage": { "type": "number" },
              "trend": { "type": "string", "enum": ["increase", "decrease", "stable"] }
            }
          },
          "revenueOverTime": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": { "type": "string" },
                "totalRevenue": { "type": "number" }
              }
            }
          },
          "topProducts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "name": { "type": "string" },
                "image": { "type": "string", "nullable": true },
                "totalCount": { "type": "integer" },
                "totalRevenue": { "type": "number" }
              }
            }
          },
          "filter": { "$ref": "#/components/schemas/StatsFilter" },
          "timeRangeLabel": { "type": "string" }
        }
      }
    },
    "parameters": {
      "OrderId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 12,
          "maxLength": 36,
          "description": "Order public id (lowercase hex or legacy UUID)."
        }
      },
      "TableId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "ZoneId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 12,
          "maxLength": 36,
          "description": "Table zone public id."
        }
      },
...
      "TenantHeaderOptional": {
        "name": "x-tenant-id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Tenant slug (e.g. demo) or registry tenant UUID. Optional when the request includes a valid staff Bearer JWT—the server resolves tenant from the token."
      },
...
    }
  },
  "paths": {
...
    "/api/v1/orders/{id}/table": {
      "patch": {
        "tags": ["Orders"],
        "summary": "Move dine-in order to another table",
        "description": "Updates **`tableId`** on an open dine-in order (`confirmed` or `preparing`). Refreshes occupancy on old and new tables. Broadcasts **`order.updated`** and **`table.updated`** on the staff WebSocket. Target table must not already have another **`confirmed`** or **`preparing`** dine-in order.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          { "$ref": "#/components/parameters/OrderId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatchOrderTableRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order": { "$ref": "#/components/schemas/OrderSerializedEnriched" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation / not dine-in / order closed / table not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Target table already occupied",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mobile/zones": {
      "get": {
        "tags": ["Tables"],
        "summary": "List zones with tables (mobile)",
        "description": "Each zone includes **`tables[]`** (active rows only, same ordering as **`GET /tables`**). Use table **`id`** as **`tableId`** on dine-in orders.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TenantHeaderOptional" }],
        "responses": {
          "200": {
            "description": "OK — `{ zones: TableZoneWithTablesMobile[] }`",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["zones"],
                  "properties": {
                    "zones": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TableZoneWithTablesMobile" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/table-zones": {
      "get": {
        "tags": ["Table zones"],
        "summary": "List table zones",
        "description": "Returns all zones ordered by **`sortOrder`**, then **`name`**. Use **`id`** as **`zoneId`** when creating or moving tables.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TenantHeaderOptional" }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["zones"],
                  "properties": {
                    "zones": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TableZone" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Table zones"],
        "summary": "Create table zone",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TenantHeaderOptional" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TableZoneCreateBody" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["zone"],
                  "properties": {
                    "zone": { "$ref": "#/components/schemas/TableZone" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Duplicate zone name (`zone_name_taken`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/table-zones/{id}": {
      "patch": {
        "tags": ["Table zones"],
        "summary": "Update table zone",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          { "$ref": "#/components/parameters/ZoneId" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TableZonePatchBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["zone"],
                  "properties": {
                    "zone": { "$ref": "#/components/schemas/TableZone" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Zone not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Duplicate zone name (`zone_name_taken`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Table zones"],
        "summary": "Delete table zone",
        "description": "Hard delete. Fails with **`409 zone_has_tables`** if any **`RestaurantTable`** still references this zone (FK **Restrict**).",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          { "$ref": "#/components/parameters/ZoneId" }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Zone not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Zone still has tables (`zone_has_tables`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tables": {
      "get": {
        "tags": [
          "Tables"
        ],
        "summary": "List active tables",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantHeaderOptional"
          }
        ]
      },
      "post": {
        "tags": [
          "Tables"
        ],
        "summary": "Create table",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantHeaderOptional"
          }
        ]
      }
    },
...
    "/api/v1/tables/{id}": {
      "patch": {
        "tags": [
          "Tables"
        ],
        "summary": "Update table",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantHeaderOptional"
          },
          {
            "$ref": "#/components/parameters/TableId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tables"
        ],
        "summary": "Soft-delete table (isActive=false)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantHeaderOptional"
          },
          {
            "$ref": "#/components/parameters/TableId"
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stats/statistics": {
      "get": {
        "tags": ["Stats"],
        "summary": "Dashboard statistics",
        "description": "Legacy-style POS statistics adapted to maklaGo orders. **`filter`**: `today` (default), `week`, `month`, `year`, or `custom`. Optional **`startDate`** / **`endDate`** (`YYYY-MM-DD`, Africa/Tunis day boundaries) override the filter range. Revenue includes non-canceled orders (`waiting`, `confirmed`, `preparing`, `completed`). Amounts in major currency units.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          {
            "name": "filter",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/StatsFilter" }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "Inclusive start (Africa/Tunis timezone)"
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "Inclusive end (Africa/Tunis timezone)"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatisticsResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/summary": {
      "get": {
        "tags": [
          "Stats"
        ],
        "summary": "Order / revenue aggregates",
        "description": "Optional `from` and `to` (ISO 8601) filter `orders.createdAt`.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantHeaderOptional"
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Inclusive lower bound on order createdAt"
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Inclusive upper bound on order createdAt"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}